开始向高阶算法进发

+++ src/main/java/leetcode/level/easy/rand5/FirstBadVersion.java
@@ -51,7 +51,7 @@ public class FirstBadVersion extends VersionControl {
         if (end - start < 32) {
             return findInRange(start, end);
         }
-        int mid = start + end / 2;
+        int mid = (int)(((long) start + (long)end) / 2);
         boolean isMidBad = isBadVersion(mid);
         if (isMidBad) {
             return findBinary(start, mid);
